Cadastro com jQuery iniciado

+++ app_users/user_controller_jq.php
@@ -3,14 +3,14 @@ require "connection.php";
 require "user_model.php";
 require "user_service.php";
 
-if ($_GET['actionUser'] == 'searchAllUsers') {
+if (isset($_GET['actionUser']) && $_GET['actionUser'] == 'searchAllUsers') {
 
     $connection = new ConnectionDB();
     $user = new User();
     $userService = new UserService($connection, $user);
     $users = $userService->readAllUsers();
     echo json_encode($users);
-} else if ($_GET['actionUser'] == 'searchTermName' && isset($_GET['name'])) {
+} else if (isset($_GET['actionUser']) &&  $_GET['actionUser'] == 'searchTermName' && isset($_GET['name'])) {
 
     $connection = new ConnectionDB();
     $user = new User();
@@ -18,11 +18,14 @@ if ($_GET['actionUser'] == 'searchAllUsers') {
     $userService = new UserService($connection, $user);
     $users = $userService->searchUsers();
     echo json_encode($users);
-} else if ($_GET['actionUser'] == 'deleteUser' && isset($_GET['userId']) && is_numeric($_GET['userId'])) {
+} else if (isset($_GET['actionUser']) && $_GET['actionUser'] == 'deleteUser' && isset($_GET['userId']) && is_numeric($_GET['userId'])) {
 
     $connection = new ConnectionDB();
     $user = new User();
     $user->__set('id', $_GET['userId']);
     $userService = new UserService($connection, $user);
     $userService->deleteUser();
+} else if ($_POST['actionUser'] == 'registerUser') {
+
+    echo json_encode($_POST);
 }
